chore: refactoring packages
