+++ cmd/crank/marketplace/login.go
@@ -36,8 +36,8 @@ import (
 
 	"github.com/crossplane/crossplane-runtime/pkg/errors"
 
-	"github.com/crossplane/crossplane/cmd/crank/internal/config"
 	"github.com/crossplane/crossplane/cmd/crank/internal/upbound"
+	"github.com/crossplane/crossplane/cmd/crank/internal/upbound/config"
 )
 
 const (
